processing/requirements.txt
flask
python-dotenv
flair
pytest
pytest-mock
fitz
